Update Chat example: rename the package name from todo to char
