+++ android/ObjectPermissionAdvanced/app/src/main/AndroidManifest.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.realm.todo">
+    package="io.realm.chat">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".ToDoApplication"
+        android:name=".ChatApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
