+++ android/ObjectPermissionAdvanced/app/src/main/java/io/realm/chat/model/PrivateChatRoom.java
@@ -1,16 +1,18 @@
-package io.realm.todo.model;
+package io.realm.chat.model;
 
 import io.realm.RealmList;
 import io.realm.annotations.PrimaryKey;
 import io.realm.annotations.RealmClass;
 import io.realm.annotations.Required;
+import io.realm.sync.permissions.Permission;
 
 @RealmClass
-public class PublicChatRoom implements ChatRoom {
-    @Required
+public class PrivateChatRoom implements ChatRoom {
     @PrimaryKey
+    @Required
     private String name;
     private RealmList<Message> messages;
+    private RealmList<Permission> permissions;
 
     @Override
     public String getName() {
@@ -20,6 +22,10 @@ public class PublicChatRoom implements ChatRoom {
     @Override
     public RealmList<Message> getMessages() {
         return messages;
+    }
+
+    public RealmList<Permission> getACL() {
+        return permissions;
     }
 
     public void setName(String name) {
