+++ android/ObjectPermissionAdvanced/app/src/main/java/io/realm/chat/model/PublicChatRoom.java
@@ -1,18 +1,16 @@
-package io.realm.todo.model;
+package io.realm.chat.model;
 
 import io.realm.RealmList;
 import io.realm.annotations.PrimaryKey;
 import io.realm.annotations.RealmClass;
 import io.realm.annotations.Required;
-import io.realm.sync.permissions.Permission;
 
 @RealmClass
-public class PrivateChatRoom implements ChatRoom {
-    @PrimaryKey
+public class PublicChatRoom implements ChatRoom {
     @Required
+    @PrimaryKey
     private String name;
     private RealmList<Message> messages;
-    private RealmList<Permission> permissions;
 
     @Override
     public String getName() {
@@ -22,10 +20,6 @@ public class PrivateChatRoom implements ChatRoom {
     @Override
     public RealmList<Message> getMessages() {
         return messages;
-    }
-
-    public RealmList<Permission> getACL() {
-        return permissions;
     }
 
     public void setName(String name) {
