+++ android/ObjectPermissionAdvanced/settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':android-adapters-debug'
+include ':app'
